Utilities for interacting with PNNL DMS

Fetches results of different data analysis tools. Not useful outside of PNNL unless connected through VPN. Works on windows out of the box. To make it work on Mac/Linux FreeTDS needs to be installed. Mac/Linux functionality is not well tested.

get_dms_job_records(): returns job records given variety keyword patterns

get_tool_output_files_for_job_number(): returns the output of the tool given tool name and file pattern

get_output_folder_for_job_and_tool(): returns the path given job id and tool name

get_AScore_results(): returns Ascore results given data package number

get_job_records_by_dataset_package(): returns job records given data package number

get_results_for_multiple_jobs(): returns concatenated results given job numbers

get_results_for_multiple_jobs.dt(): returns results as concatenated data.table given job numbers

get_results_for_single_job(): returns results given job number

get_results_for_single_job.dt(): returns results as data.table given job number

get_datasets_by_data_package(): returns table with datasets info

download_datasets_by_data_package(): downloads datasets into provided location

tool2suffix

is_PNNL_DMS_connection_successful()

get_dms_job_records(
  jobs = NULL,
  datasetPttrn = "",
  experimentPttrn = "",
  toolPttrn = "",
  parPttrn = "",
  settingsPttrn = "",
  fastaPttrn = "",
  proteinOptionsPttrn = "",
  instrumentPttrn = ""
)

get_tool_output_files_for_job_number(
  jobNumber,
  toolName = NULL,
  fileNamePttrn,
  mostRecent = TRUE
)

get_output_folder_for_job_and_tool(jobNumber, toolName, mostRecent = TRUE)

get_job_records_by_dataset_package(data_package_num)

get_datasets_by_data_package(data_package_num)

download_datasets_by_data_package(
  data_package_num,
  copy_to = ".",
  fileNamePttrn = ".raw",
  ncores = 2
)

get_results_for_multiple_jobs(jobRecords)

get_results_for_multiple_jobs.dt(jobRecords, expected_multiple_files = FALSE)

get_results_for_single_job(pathToFile, fileNamePttrn)

get_url_from_dir_and_file(dir, file_name_segment)

get_results_for_single_job.dt(
  pathToFile,
  fileNamePttrn,
  expected_multiple_files = FALSE
)

path_to_FASTA_used_by_DMS(data_package_num, organism_db = NULL)

Format

An object of class list of length 7.

Arguments

jobs

(integer) DMS job ID

datasetPttrn

experimentPttrn

toolPttrn

parPttrn

settingsPttrn

fastaPttrn

proteinOptionsPttrn

instrumentPttrn

jobNumber

toolName

fileNamePttrn

mostRecent

(logical) only most recent or all output files

data_package_num

(integer) data package ID number.

copy_to

folder path to copy files into

ncores

number of cores to use in cluster

jobRecords

expected_multiple_files

do we expect multiple files for a single job or not. Default is FALSE.

pathToFile

dir

file_name_segment

organism_db

(character) FASTA file. This is the same as the Organism DB column. No need to specify this if there is only one FASTA file associated with the jobs.

Examples

if (is_PNNL_DMS_connection_successful()) {
  get_output_folder_for_job_and_tool(863951, "DTA_Refinery")
}