Reading MS-GF+ output from PNNL's DMS

Reading MS-GF+ output from PNNL's DMS

read_msgf_data_from_DMS(
  data_package_num,
  param_file = NULL,
  organism_db = NULL,
  use_mzIdentML = FALSE
)

Arguments

data_package_num

numeric or character; the Data Package ID in the DMS.

param_file

character; MS-GF+ parameter file. No need to specify this if there is only one parameter file associated with the jobs.

organism_db

character; FASTA file name. This is the same as the Organism DB column. No need to specify this if there is only one FASTA file associated with the jobs.

use_mzIdentML

logical; whether to read mzIdentML files into psms data.table slot of the MSnID.

Value

(MSnID) MSnID object

Details

If use_mzIdentML = TRUE, a progress bar will not be displayed while the results are being fetched. Also, the table in the psms slot will not be the same as when mzIdentML = FALSE. As a result, it may require additional modification by the user to prepare it for downstream processing steps.

Examples

if (is_PNNL_DMS_connection_successful()) {
  msnid <- read_msgf_data_from_DMS(3606)
  show(msnid)

  # msnid <- read_msgf_data_from_DMS(3606, use_mzIdentML = TRUE)
  # show(msnid)
}