Reading and Writing Study Design Tables

Fetches study design results from DMS. Checks that study design files are internally consistent. DMS functionality not useful outside of PNNL unless connected through VPN.

path_to_study_design_from_DMS(): gets location of study design files from the DMS.

read_study_design_from_DMS(): reads study design tables from the DMS data package folder.

write_study_design_to_DMS(): writes study design tables to the DMS so that they can be accessed by others.

path_to_study_design_from_DMS(data_package_num)

read_study_design_from_DMS(data_package_num, useHTTP = FALSE)

write_study_design_to_DMS(data_package_num, study_design, overwrite = FALSE)

Arguments

data_package_num

(integer) data package number for DMS

useHTTP

(logical) whether to...

study_design

(list) study design files: fractions, samples, and references.

overwrite

(logical) whether to replace any existing study design tables when writing files.

Examples

if (is_PNNL_DMS_connection_successful()) {
  study_design <- read_study_design_from_DMS(3606)

  fractions  <- study_design$fractions
  samples    <- study_design$samples
  references <- study_design$references
}